refactor: move enumerateDistinctFluids() to PumpCover
